requirements.txt
# Core requirements
pandas==2.0.0
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1

# Data visualization
matplotlib==3.7.1
seaborn==0.12.2

# Testing
pytest==7.3.1
